move the logic to a script
